// src/api/v1/ollama/dto/ollama.response.ts
export class OllamaResponse {
  model: string;
  created_at: Date;
  response: string;
  done: boolean;
  done_reason: string;
  context: number[];
  total_duration: number;
  load_duration: number;
  prompt_eval_count: number;
  prompt_eval_duration: number;
  eval_count: number;
  eval_duration: number;
  duration_ms: number;
}